backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
python-multipart
supabase
httpx
pandas
python-dotenv
pypdf
python-pptx
unstructured
openpyxl
google-generativeai
langchain-text-splitters
